Generalize the ExtensionFunctionDispatcher::Delegate interface a bit more. In particular remove GetExtensionHost() and GetExtensionDOMUI(). I'm going to add a new type of EFD::Delegate soon that is neither of these, and I don't think it makes sense to have the code assume that EFD::Delegate must be one of the two of them.
Some code still does RTTI stuff through RenderViewHostDelegate::GetRenderViewType(), but in that case it is more clear that there are a long list of potential view types, and that the caller must be more careful.

Review URL: http://codereview.chromium.org/1149003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@42519 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 0ec9203dac6b11c664adda8a39b77f3f00eb6072

+++ chrome/browser/extensions/extension_bookmark_manager_api.cc
@@ -21,6 +21,7 @@
 #include "chrome/browser/importer/importer.h"
 #include "chrome/browser/importer/importer_data_types.h"
 #include "chrome/browser/profile.h"
+#include "chrome/browser/renderer_host/render_view_host.h"
 #include "chrome/browser/tab_contents/tab_contents.h"
 #include "grit/generated_resources.h"
 
@@ -396,10 +397,18 @@ bool StartDragBookmarkManagerFunction::RunImpl() {
   EXTENSION_FUNCTION_VALIDATE(
       GetNodesFromArguments(model, args_as_list(), &nodes));
 
-  bookmark_utils::DragBookmarks(profile(), nodes,
-      dispatcher()->GetExtensionDOMUI()->tab_contents()->GetNativeView());
+  if (dispatcher()->render_view_host()->delegate()->GetRenderViewType() ==
+      ViewType::TAB_CONTENTS) {
+    ExtensionDOMUI* dom_ui =
+        static_cast<ExtensionDOMUI*>(dispatcher()->delegate());
+    bookmark_utils::DragBookmarks(
+        profile(), nodes, dom_ui->tab_contents()->GetNativeView());
 
-  return true;
+    return true;
+  } else {
+    NOTREACHED();
+    return false;
+  }
 }
 
 bool DropBookmarkManagerFunction::RunImpl() {
@@ -432,15 +441,24 @@ bool DropBookmarkManagerFunction::RunImpl() {
   else
     drop_index = drop_parent->GetChildCount();
 
-  ExtensionBookmarkManagerEventRouter* router = dispatcher()->
-      GetExtensionDOMUI()->extension_bookmark_manager_event_router();
+  if (dispatcher()->render_view_host()->delegate()->GetRenderViewType() ==
+      ViewType::TAB_CONTENTS) {
+    ExtensionDOMUI* dom_ui =
+        static_cast<ExtensionDOMUI*>(dispatcher()->delegate());
+    ExtensionBookmarkManagerEventRouter* router =
+        dom_ui->extension_bookmark_manager_event_router();
 
-  DCHECK(router);
+    DCHECK(router);
 
-  bookmark_utils::PerformBookmarkDrop(profile(), *router->GetBookmarkDragData(),
-                                      drop_parent, drop_index);
+    bookmark_utils::PerformBookmarkDrop(profile(),
+                                        *router->GetBookmarkDragData(),
+                                        drop_parent, drop_index);
 
-  router->ClearBookmarkDragData();
-  SendResponse(true);
-  return true;
+    router->ClearBookmarkDragData();
+    SendResponse(true);
+    return true;
+  } else {
+    NOTREACHED();
+    return false;
+  }
 }
